move video poster under video preview
